PluginBrowser.py: change internet checking method for feeds
This commit changes the internet checking in C enigma checkInternetAccess
to a more conventional method like resolving a URL with request and urlopen.
These methods are also reliable but have a shorter response time and this response time is always when accessing the browser plugin and never once within this screen as chekInternetAccess does. Once inside the browser plugin, ChekInternetAccess can take up to three seconds to receive the response and this time slows down navigation through this screen until it receives response zero.